pyproject.toml
[project]
name = "devops-engineer-from-scratch-project-313"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "flask>=3.1.2",
    "gunicorn>=24.1.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sentry-sdk[flask]>=2.51.0",
    "sqlmodel>=0.0.31",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-flask>=1.3.0",
    "ruff>=0.14.14",
]
